Python/SDF turned all vectors into real

+++ python/sdf.py
@@ -83,7 +83,7 @@ class SDF:
 
     def load_mp(self, iscomplex):
         [ self.load_tensor(iscomplex) for i in range(sefl.read_longs(1))]
-    
+
     def load_tensor(self, iscomplex):
         rank = self.read_longs(1)
         dims = self.read_longs(rank)
@@ -92,7 +92,7 @@ class SDF:
             data = self.read_complex(L)
         else:
             data = self.read_doubles(L)
-        return np.ndarray(shape=dims,buffer=data,order='F')
+        return np.ndarray(shape=dims,buffer=data,dtype=data.dtype,order='F')
 
     def read_longs(self, n):
         output = np.ndarray(shape=(n,),dtype=self.long_type)
